Extract TemplateNavigation
